Removed Microsoft.Extensions.DependencyModel from net451
