+++ src/DotNetToolkit.Repository.AdoNet/Internal/DataAccessProviderType.cs
@@ -1,0 +1,15 @@
+﻿namespace DotNetToolkit.Repository.AdoNet.Internal
+{
+    internal enum DataAccessProviderType
+    {
+        SqlServer,
+        SqLite,
+        MySql,
+        PostgreSql,
+
+#if NETFULL
+        OleDb,
+        SqlServerCompact
+#endif
+    }
+}
